requirements.txt
streamlit==1.39.0   
mysql-connector-python==9.0.0   
bcrypt==4.2.0              
cryptography==43.0.1        
secrets                     
base64                     
pandas==2.2.3              
numpy==2.1.2                
re                         
string                     
python-dotenv==1.0.1      
